// src/app/model/ambtracking.ts
export interface Ambtracking {
    ambassadorid: number;
    created: string;
    extrainfo: string;
    id: number;
    jobdetailsId: number;
    jobdetailsid: number;
    jobscheduledate: string;
    jobscheduleid: number;
    lastupdated: string;
    loginimageurl: string;
    loginlocation: string;
    logintime: string;
    logoutimageurl: string;
    logoutlocation: string;
    logouttime: string;
    published: number;
    sessiontimesid: number;
    setupimageurl: string;
    setuplocation: string;
    setuptime: string;
    trackingstatus: number;
    venueid: number;
}